refactoring blacklist service

+++ src/main/java/Blacklist/Manager/repository/BlacklistLogRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlacklistLogRepository extends JpaRepository<BlacklistLog, Integer> {
-    // Define custom query methods here if needed
-}
+public interface BlacklistLogRepository extends JpaRepository<BlacklistLog, Long> {}
+
